consumer/requirements.txt
boto3==1.21.18
botocore==1.24.18
pika==1.2.0
Pillow==9.0.1
psutil==5.9.0
psycopg2==2.9.3
requests==2.27.1